requirements.txt
certifi==2020.12.5
chardet==4.0.0
fake-useragent==1.1.3
idna==2.10
PyJWT==1.7.1
pytz==2021.1
PyYAML==6.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
twilio==6.55.0
urllib3==1.26.4
